Switch arithmetic to long long with overflow checks

+++ src/arith.h
@@ -1,6 +1,6 @@
 #ifndef ARITH_H
 #define ARITH_H
 
-long eval_arith(const char *expr, int *err);
+long long eval_arith(const char *expr, int *err);
 
 #endif /* ARITH_H */
